Move the support classes for work with the local db to domain.room package
