+++ app/src/main/java/ru/jjba/jr2/data/db/dao/SectionDao.kt
@@ -3,7 +3,7 @@ package ru.jjba.jr2.data.db.dao
 import androidx.room.Dao
 import androidx.room.Query
 import ru.jjba.jr2.domain.entity.Section
-import ru.jjba.jr2.domain.select.SectionWithGroups
+import ru.jjba.jr2.domain.room.select.SectionWithGroups
 
 @Dao
 abstract class SectionDao : BaseDao<Section> {
