+++ app/src/main/java/ru/jjba/jr2/data/repository/WordDbRepository.kt
@@ -5,7 +5,7 @@ import ru.jjba.jr2.App
 import ru.jjba.jr2.data.db.AppDatabase
 import ru.jjba.jr2.data.db.dao.WordDao
 import ru.jjba.jr2.domain.entity.Word
-import ru.jjba.jr2.domain.join.GroupOfWordsJoin
+import ru.jjba.jr2.domain.room.join.GroupOfWordsJoin
 
 class WordDbRepository(
         db: AppDatabase = App.instance.db
