+++ app/src/main/java/ru/jjba/jr2/domain/room/join/GroupOfWordsJoin.kt
@@ -1,24 +1,24 @@
-package ru.jjba.jr2.domain.join
+package ru.jjba.jr2.domain.room.join
 
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
 import ru.jjba.jr2.domain.entity.Group
-import ru.jjba.jr2.domain.entity.Moji
+import ru.jjba.jr2.domain.entity.Word
 
 @Entity(foreignKeys = [
     ForeignKey(entity = Group::class,
             parentColumns = ["id"],
             childColumns = ["groupId"]),
-    ForeignKey(entity = Moji::class,
+    ForeignKey(entity = Word::class,
             parentColumns = ["id"],
-            childColumns = ["mojiId"])
+            childColumns = ["wordId"])
 ])
-class GroupOfKanjisJoin(
+class GroupOfWordsJoin(
         @PrimaryKey(autoGenerate = true)
-        val id: Int,
+        val id: Int = 0,
         val groupId: Int,
-        val kanjiId: Int,
-        val defaultPossition: Int,
-        val customPossition: Int
+        val wordId: Int,
+        val defaultPosition: Int = 0,
+        val customPosition: Int = 0
 )
