+++ app/src/main/java/ru/jjba/jr2/presentation/ui/vocab/word/group/WordGroupSectionViewHolder.kt
@@ -4,7 +4,7 @@ import android.view.View
 import kotlinx.android.synthetic.main.item_word_group_section.view.*
 import org.jetbrains.anko.sdk25.coroutines.onClick
 import org.zakariya.stickyheaders.SectioningAdapter
-import ru.jjba.jr2.domain.select.SectionWithGroups
+import ru.jjba.jr2.domain.room.select.SectionWithGroups
 
 class WordGroupSectionViewHolder(
         private val view: View,
